added functions

+++ philo_one/philo_one.h
@@ -6,7 +6,7 @@
 /*   By: amoussai <amoussai@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/20 10:34:08 by amoussai          #+#    #+#             */
-/*   Updated: 2021/03/21 18:58:09 by amoussai         ###   ########.fr       */
+/*   Updated: 2021/04/10 14:45:24 by amoussai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,11 @@
 #define PHILO_ONE_H
 
 #include <pthread.h> 
+#include <stdlib.h>
 #include <sys/time.h>
 #include <unistd.h>
+#include <limits.h>
+#include <stdio.h>
 
 typedef struct		s_phil
 {
@@ -45,7 +48,7 @@ typedef struct		s_state
 }					t_state;
 
 int		ft_atoi(const char *str);
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long n, int fd);
 void	ft_putstr_fd(char *s, int fd);
 
 #endif
